+++ test/utils_request.test.js
@@ -104,25 +104,28 @@ describe('Apify.utils_request', () => {
         });
 
         const testQueries = [
-            ['abc', 'abc'],
-            ['%20', ' '],
-            ['%cf', '%cf'],
-            ['helios-–-the-primordial-sun', 'helios-–-the-primordial-sun'],
-            ['helios-%E2%80%93-the-primordial-sun', 'helios-–-the-primordial-sun'],
+            ['abc', 'abc', true],
+            ['abc', 'abc', false],
+            ['%20', ' ', false],
+            ['%cf', '%cf', true],
+            ['helios-–-the-primordial-sun', 'helios-–-the-primordial-sun', true],
+            ['helios-%E2%80%93-the-primordial-sun', 'helios-–-the-primordial-sun', false],
         ];
 
-        test.each(testQueries)(`it works with not encoded urls: '%s' (regular)`, async (query, decoded) => {
+        test.each(testQueries)(`it works with not encoded urls: '%s' (regular)`, async (query, decoded, forceEncode) => {
             const response = await requestAsBrowser({
                 url: `http://${HOSTNAME}:${port}/test-encoding/?q=${query}`,
+                forceUrlEncoding: forceEncode,
             });
             expect(response.statusCode).toBe(200);
             expect(JSON.parse(response.body).q).toBe(decoded);
         });
 
-        test.each(testQueries)(`it works with not encoded urls: '%s' (stream)`, async (query, decoded) => {
+        test.each(testQueries)(`it works with not encoded urls: '%s' (stream)`, async (query, decoded, forceEncode) => {
             const response = await requestAsBrowser({
                 url: `http://${HOSTNAME}:${port}/test-encoding/?q=${query}`,
                 stream: true,
+                forceUrlEncoding: forceEncode,
             });
 
             const chunks = [];